+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.1.0
+
+- :boom: **BREAKING CHANGES** - Kuliyyah parameter now is seperated from the albiruni constructor.
+  - Before:
+  ```dart
+  var albiruni = Albiruni(kulliyah: "AED", semester: 2, session: "2021/2022");
+  var subjects = await albiruni.fetch();
+  ```
+  - After:
+  ```dart
+  var albiruni = Albiruni(semester: 2, session: "2021/2022");
+  var subjects = await albiruni.fetch("AED");
+  ```
+
+
 ## 1.0.2
 
 - :see_no_evil: When fetching subject, you now can set **`useProxy`** to `true`. This will proxy the connection from albiruni origin site. Useful when dealing with CORS error in web application.